+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.6] - Ongoing
+## [1.0.6] - 2026-08-17
 
-This is an ongoing update
+This update introduces a minor change
+
+### Changed
+ - Automated reports now omit scan-rule entries with a zero percentage in the generated message, retaining only rules
+   that affected the calculated result.